INFRA-7949: Initial commit
We currently have notifications when we recycle EC2 instances which
provide useful information when e.g. debugging an unexpected issue.
Over the last couple of years we have containerised a lot of services
that used to be AMI based and now we do not know when e.g. a new version
of an image deploys.
We'd like to restore deployment notifications to our containerised
services. This Lambda is designed to consume EventBridge ECS service
deployment events and produce Slack notifications.

Add initial code and scaffolding for
ecs-service-deployment-notifications Lambda.