INFRA-7949: Simplify sample event fixtures
The sample events fixtures are examples of JSON payloads that the Lambda
might receive from EventBridge. In an attempt to reduce duplication
these events were built from multiple dictionary fragments but this made
them hard to read and understand.

Rework sample events to be simpler and update tests to work with the new
events.